% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showLength.R
\name{showLength}
\alias{showLength}
\title{Draw the length distribution of the circRNA according to the given bed file
User can customize the upper limit of the length}
\usage{
showLength(bed, max_length)
}
\arguments{
\item{bed}{bed file}

\item{max_length}{the max length of circRNA showed in plot}
}
\description{
Draw the length distribution of the circRNA according to the given bed file
User can customize the upper limit of the length
}
